Add test for large_files
The new fixture is primarily intended to be used as a benchmark
(like the linux test), and is thus ignored by default. dev_utils
was updated to fill files with random data generated using the rand
crate instead of it being read from /dev/random, which makes hydrating
large files *much* faster.