Prevent files/directories from being copied onto/into themselves
The essence of this change is issuing a descriptive error message and
immediately aborting whenever the source(s) provided to fcp alias with
the destination.

Previously fcp would behave poorly in these scenarios, truncating the
file in the case of a normal file, and attempting to copy recursively
until the maximum path length was exceeded in the case of a directory.

A battery of test-cases has been added to ensure that fcp detects
aliasing, even in edge-cases where the aliasing is indirect (e.g.
copying a directory's parent into itself, aliasing via a symlink, etc.)