Refactor error-handling utilities
The Error type and its associated methods have been moved into a
separate module, since it is used outside of just the filesystem module
(where it was previously defined). Error has been rewritten as a
tuple struct to better reflect that is is a simple newtype wrapper.

In addition to the above, a custom Result type has been introduced
(similar to std::io::Result) which uses Error as its Err type,
cutting down on boilerplate.