Improve performance by avoiding unnecessary `statx`/`fstat64` and `fchmodat` syscalls
While profiling `fcp` on Linux, I happened to notice that there were
significantly more `statx` syscalls being made than I would have
expected, approximately 2x the number of files being copied, instead of
just 1x as expected. It turns out the culprit was the Rust standard
library implementation of `std::fs::copy`, which guards against some
scenarios we don't care about in the context of `fcp`, but cost
additional syscalls.

This has been addressed. On Linux, this results in a 50% reduction in
`statx` syscalls and a 100% reduction in `fchmodat` syscalls. On macOS,
the reductions are the same (with `fstat64` in place of `statx`), except
that they only are realized when copying between two different
filesystems, or within a single filesystem that does not support
`fclonefileat` (when `fclonefileat` is available the aforementioned
syscalls don't need to be made at all).